Fixed Issue 85: POST request failed Fixed Issue 87: BUG Save Fixed Issue 91: Auto-complete prevents entering different URI Added Magic Variables feature Added qearch to the saved requests ad history view Added delete history entry button to the history view Some changes in core data layer Compiled with a latest stable GWT version (2.5.0)
This version has been released in chrome web store.

git-svn-id: https://chrome-rest-client.googlecode.com/svn/trunk@179 4310ba71-df81-c583-98c1-67588e199800

Former-commit-id: c0921566282667bb1e6bd591ad38af32d37f8b40 [formerly 39104edeeb3629799fcfe4533858ba1c272ec5da]
Former-commit-id: 7a2e3538f0f2fc7e4e8d66331085e5a7af5ccf2b

+++ RestClient/src/org/rest/client/storage/IndexedDbAdapter.java
@@ -388,11 +388,11 @@ public abstract class IndexedDbAdapter<K, V extends JavaScriptObject>
 
 			IDBObjectStore<K> store = (IDBObjectStore<K>) tx
 					.objectStore(storeName);
-			final IDBRequest<Long> request = store.count(key);
+			final IDBRequest<Integer> request = store.count(key);
 			request.addSuccessHandler(new IDBSuccessHandler() {
 				@Override
 				public void onSuccess() {
-					Long r = request.getResult();
+					Integer r = request.getResult();
 					if (r != null && r > 0) {
 						callback.onSuccess(true);
 					} else {
@@ -541,13 +541,13 @@ public abstract class IndexedDbAdapter<K, V extends JavaScriptObject>
 
 	@SuppressWarnings("unchecked")
 	@Override
-	public void countAll(final StoreResultCallback<Long> callback) {
+	public void countAll(final StoreResultCallback<Integer> callback) {
 		try {
 			IDBTransaction tx = db.transaction(storeName,
 					IDBTransaction.READ_ONLY_TRANSACTION);
 			final IDBObjectStore<String> store = (IDBObjectStore<String>) tx
 					.objectStore(storeName);
-			final IDBRequest<Long> request = store.count();
+			final IDBRequest<Integer> request = store.count();
 			request.addSuccessHandler(new IDBSuccessHandler() {
 				@Override
 				public void onSuccess() {
