+++ RestClient/src/org/rest/client/storage/store/HeadersStoreWebSql.java
@@ -90,7 +90,7 @@ public class HeadersStoreWebSql extends WebSqlAdapter<Integer, HeaderRow> {
 	 * This type of query is not available for this table.
 	 */
 	@Override
-	public void countAll(StoreResultCallback<Long> callback) {
+	public void countAll(StoreResultCallback<Integer> callback) {
 		callback.onError(null);
 	}
 	/**
